chore: Work on minimizing UMD bundle size
- Always use LocalOutlet in client rather than normal Outlet.
- Create AbstractOutlet to share methods required by both LocalOutlet & Outlet.
- Dynamically load several classes when required:
  - LocalOutlet
  - Web3WalletProvider (this was already being dynamically import but also statically
    imported for type usage. I've updated this to import type only)
  - Authentication methods ticketZKProof & signedUNChallenge
  - Authenticator library
- Cleanup and remove some redundant functionality in Outlet.
